cast to avoid bitloss
git-svn-id: https://svn.code.sf.net/p/bzflag/code/branches/v1_7/bzflag@581 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzfs/bzfs.cxx
@@ -382,7 +382,8 @@ private:
     else
       b[3] = atoi(ip);
 
-	mask.s_addr= b[0] << 24 | b[1] << 16 | b[2] << 8 | b[3];
+	mask.s_addr= ((unsigned int)b[0] << 24) |
+        ((unsigned int)b[1] << 16) | ((unsigned int)b[2] << 8) | b[3];
     return true;
   }
